hodina

+++ promenne/Povrch a objem valce.py
@@ -1,7 +1,7 @@
 import math
 
-height = int(input("height "))
-radius = int(input("radius "))
+height = float(input("height "))
+radius = float(input("radius "))
 volume = math.pi * radius ** 2 * height
 print(f"{volume} cm³")
 surface = 2 * math.pi * radius * (radius + height)
