+++ promenne/Prepocet teploty.py
@@ -1,6 +1,6 @@
 import math
 
-temperature_celsius = int(input("temperature [°C]"))
+temperature_celsius = float(input("temperature [°C]"))
 temperature_kelvin = temperature_celsius * 273.15
 temperature_farenheit = temperature_celsius * 9/5 + 32
